Report the distances, spending and credit change the journal states
A jump distance is stated as a float and the magnitude reader took integers only, so every distance was discarded and a session of five jumps reported nought light years. Accept both, and widen the magnitude, the flight rollup and the distance formatter to match.

The net credit change was folded from the moments, which prices income alone, so a session paying an eleven million credit rebuy still reported a gain. It is now the difference between the balances the journal states at each end, and is unread rather than zero when fewer than two are stated.

A purchase cost is stated in TotalCost but reported as nought spent. Route it through the magnitude channel so it is read without folding spending into an income total.

Derive the fleet carrier distance from consecutive CarrierJump positions, carrying the measured leg count so an unmeasurable first leg is declared rather than hidden.

Watch magnitude fields in the field diagnostic: the jump defect was invisible to it because the field was present and merely unread. Delete the aggregate config blocks, which nothing has ever parsed.

+++ o7debrief/domain/model/conceptual_moment.py
@@ -31,7 +31,7 @@ class ConceptualMoment:
     mode: ActivityMode
     occurred_at: EventTime
     label: str
-    magnitude: int
+    magnitude: float
     credits_delta: Credits
     detail: tuple[tuple[str, object], ...]
     # A second, distinct currency delta (Operations pay Merc Coins). It rides
